feat: bull share feature ready
